refactor: rename files to support new examples of flatlist
