+++ e2e/flatListGrid/scrollList.yaml
@@ -1,7 +1,7 @@
 appId: com.ronicesarrc.performance.demo
 ---
 - repeat:
-    times: 10 # Repeat enough times to scroll to the bottom
+    times: 5 # Repeat enough times to scroll to the bottom
     commands:
       - swipe:
           direction: UP
